Remove special command for downdloading self-response rate
I reworked some of my customized versions of the core downloader
classes to handle field names and types that vary from year to year,
as they do with the response rate endpoints.

I also implemented a base table class for non-ACS tables that
uses the downloader class (right now, for tracts only). This allows
hooking into my custom version of the downloader CLI command instead of
having to implement a second command.

This seems like it will be a useful step forward for bringing all of my
custom downloads into one command and for thinking of how to make the
mainline census-data-downloader package more extensible.